% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afni.R
\name{afni}
\alias{afni}
\title{A wrapper for issuing AFNI commands to terminal.}
\usage{
afni(.fun, .args, afni.path, ...)
}
\arguments{
\item{.fun}{AFNI function to call}

\item{.args}{a character vector of length 1 with all commands following function call}

\item{afni.path}{Optional; location of AFNI on machine. If not specified, will guess path based on `Sys.info()["nodename"]`}
}
\description{
Depends: If on a windows platform, will require WSL.
}
\keyword{AFNI,}
\keyword{neuroimaging}
