% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_overlay.R
\name{build_overlay}
\alias{build_overlay}
\title{Plot statistics on a surface.}
\usage{
build_overlay(
  d,
  col.values,
  col.roi.num = "num.roi",
  col.hemi = "hemi",
  template = mmp
)
}
\arguments{
\item{d}{data.frame with one row per ROI.}

\item{col.values}{name for column of values for overlay}

\item{col.roi.num}{name for column of roi numbers to match in template}

\item{col.hemi}{name for column indicating hemisphere}

\item{template}{which underlay to use?}
}
\description{
Given an overlay (e.g., statistics) and underlay (e.g., group-template gifti file).
Overlay and underlay must have same structure: lists of length two, with names "L" and "R".
Elements within .$L and .$R correspond to vertices.
}
\keyword{AFNI,}
\keyword{data}
\keyword{gifti,}
\keyword{neuroimaging,}
\keyword{vis}
