% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collate_surface_params.R
\name{collate_surface_params}
\alias{collate_surface_params}
\title{Common vector math.}
\usage{
collate_surface_params(a)
}
\arguments{
\item{name}{name of .gii file to read}

\item{xlabels}{optional; a character vector that indicates the exact strings of the sub-bricks to be extracted. If omitted, all sub-bricks are extracted.}

\item{space}{the surface space the .gii is in. \%in\% c("hcp", "fsave")}
}
\description{
Common vector math.
}
\keyword{algebra}
\keyword{centering,}
\keyword{linear}
\keyword{normalizing,}
\keyword{scaling,}
