% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combopaste.R
\name{combo.paste}
\alias{combo.paste}
\title{Paste all combinations of elements of a set of strings.}
\usage{
combo.paste(a, b, sep = ".")
}
\arguments{
\item{a}{(character) vector 1, left-hand side of output}

\item{b}{(character) vector 2, right-hand side of output}

\item{sep}{a character string to separate the results}
}
\description{
Paste all combinations of elements of a set of strings.
}
\examples{
combo.paste(letters[1:2], letters[3:4])  ##
combo.paste(letters[1:2], letters[3:4], sep = "_")
combo_paste(letters[1:2], letters[3:4])  ## some common options for sep can be specified via the the function name
combopaste3(letters[1:2], letters[3:4], letters[24:26])  ## accommodates upt to three vectors
}
\keyword{concatenating}
\keyword{manipulation,}
\keyword{string}
\keyword{strings}
