% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_matrix.R
\name{contrast_matrix}
\alias{contrast_matrix}
\title{Generates contrast matrix for forming cross-validated euclidean distance matrix.}
\usage{
contrast_matrix(n, condition.names)
}
\arguments{
\item{n}{number of conditions}

\item{condition.names}{optional}
}
\description{
Generates contrast matrix for forming cross-validated euclidean distance matrix.
}
