% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvdist.R
\name{cvdist}
\alias{cvdist}
\title{Computes cross-validated Mahalanobis distance.}
\usage{
cvdist(x, m = NULL, condition.names = NULL, W = NULL)
}
\arguments{
\item{x}{a data matrix, conditions as columns}

\item{m}{optional; a contrast matrix, e.g., created with `contrast_matrix()`}

\item{W}{optional; a whitening matrix}
}
\description{
Computes cross-validated Mahalanobis distance.
}
\keyword{RSA}
