% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot2dash.R
\name{dot2dash}
\alias{dot2dash}
\title{Substitute common characters.}
\usage{
dot2dash(x)
}
\arguments{
\item{x}{character string}
}
\description{
A wrapper for common string substitutions.
}
\keyword{gsub}
\keyword{manipulation,}
\keyword{string}
