% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt_mat.R
\name{melt_mat}
\alias{melt_mat}
\title{Reshapes matrix to long-form dataframe for use with `ggplot2()`.}
\usage{
melt_mat(R, var.names = c("v1", "v2"), val.name = "value")
}
\arguments{
\item{R}{matrix}
}
\description{
Reshapes matrix to long-form dataframe for use with `ggplot2()`.
}
\keyword{RSA}
