% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_matrix_operations.R
\name{oneminus}
\alias{oneminus}
\title{Perform common operations on (dis)similarity matrices.}
\usage{
oneminus(x)
}
\arguments{
\item{m}{a correlation matrix}
}
\description{
A wrapper for the corrplot::corrplot() function with convenient defaults.
}
\keyword{correlation}
\keyword{matrices,}
\keyword{plotting}
