% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outpro.R
\name{outpro}
\alias{outpro}
\title{Identifies multivariate (bivariate) outliers via stahel-donoho projection procedure.}
\usage{
outpro(
  m,
  gval = NA,
  center = NA,
  plotit = TRUE,
  op = TRUE,
  MM = FALSE,
  cop = 3,
  xlab = "VAR 1",
  ylab = "VAR 2",
  STAND = TRUE,
  tr = 0.2,
  q = 0.5,
  pr = TRUE,
  ...
)
}
\arguments{
\item{m}{matrix}
}
\description{
From creators of WRS2; see https://github.com/cran/WRS2/blob/master/R/outpro.R
}
