% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_surfaces.R
\name{plot_surface}
\alias{plot_surface}
\title{Plotting GIFTI surface files.}
\usage{
plot_surface(
  overlay,
  underlay,
  facet.names = c("left_lateral", "left_medial", "right_lateral", "right_medial"),
  dims = c(1, 4),
  facet.params
)
}
\arguments{
\item{overlay}{values (colors) to plot; same structure as underlay}

\item{underlay}{a list of length 2, with elements titled c("L", "R"), corresponding to each hemisphere of a gifti of the desired underlay.}

\item{facet.names}{character vector with elements \%in\% c("left_lateral", "left_medial", "right_lateral", "right_medial"). Determines the row-wise ordering of the plot matrix.}

\item{dims}{Dimensions of plot matrix (row, col).}

\item{facet.params}{Optional list, of same structure as facet.names, containing theta and phi elements which specify orientations of each facet.}
}
\description{
Much of this was adapted from code written by Joset Etzel
(http://mvpa.blogspot.com/2018/06/tutorial-plotting-gifti-images-in-r.html) and by John Muscheli (from library(gifti)).
}
\keyword{AFNI,}
\keyword{gifti}
\keyword{neuroimaging,}
