% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcor.R
\name{qcor}
\alias{qcor}
\title{Quickly plot a correlation matrix.}
\usage{
qcor(m, .title = "", ...)
}
\arguments{
\item{m}{a correlation matrix}
}
\description{
A wrapper for the corrplot::corrplot() function with convenient defaults.
}
\keyword{correlation}
\keyword{matrices,}
\keyword{plotting}
