% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_atlas.R
\name{read_atlas}
\alias{read_atlas}
\title{Reads parcellation atlas and key.}
\usage{
read_atlas(name.atlas, space = "hcp", surface = TRUE, path.atlas)
}
\description{
Reads parcellation atlas and key.
}
\keyword{AFNI,}
\keyword{neuroimaging}
