% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dmcc_behav.R
\name{read_dmcc_behav}
\alias{read_dmcc_behav}
\title{Reads trial-level behavioral data and event times for DMCC2 tasks.}
\usage{
read_dmcc_behav(
  files = "all",
  use.boxr = if (Sys.info()["nodename"] == "CCP-FREUND") TRUE else FALSE,
  boxdrive.path = "C:/Users/mcf/Box"
)
}
\arguments{
\item{files}{Character; which files to read? Must be either \%in\% "axcpt", "cuedts", "stern", "stroop", "subjsum"; or "all." Defaults to "all".}

\item{use.boxr}{Boolean; use boxr or boxdrive? When on freund's laptop, defaults to boxdrive, otherwise boxr.}

\item{boxdrive.path}{String indicating path to boxdrive. Defaults to path for freund's laptop.}
}
\description{
Reads trial-level behavioral data and event times for DMCC2 tasks.
}
\keyword{behavior}
\keyword{events,}
\keyword{read,}
