% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gifti2matrix.R
\name{read_gifti2matrix}
\alias{read_gifti2matrix}
\title{Reads a gifti image and reshapes to matrix.}
\usage{
read_gifti2matrix(name)
}
\arguments{
\item{name}{name of .gii file to read}
}
\description{
A short wrapper for gifti::read_gifti().
}
\keyword{AFNI,}
\keyword{gifti}
\keyword{neuroimaging,}
