% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xmat.R
\name{read_xmat}
\alias{read_xmat}
\title{Read an AFNI gifti image and reshapes to matrix.}
\usage{
read_xmat(name, uncensored = TRUE)
}
\arguments{
\item{name}{name of .xmat.1D file to read}

\item{uncensored}{boolean: read an uncensored file? defaults to TRUE, so nrows corresponds to other GLM matrices (Y, E).}
}
\description{
Optionally extracts sub-bricks by labels.
}
\keyword{AFNI,}
\keyword{neuroimaging}
