% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectormath.R
\name{ssq}
\alias{ssq}
\title{Common vector math.}
\usage{
ssq(v)
}
\arguments{
\item{v}{a numeric vector}

\item{m}{a numeric vector or matrix}
}
\description{
Common vector math.
}
\keyword{algebra}
\keyword{centering,}
\keyword{linear}
\keyword{normalizing,}
\keyword{scaling,}
