% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_boot_ci.R
\name{stat_boot_ci}
\alias{stat_boot_ci}
\title{Bootstrapping bivariate credible intervals for a ggplot}
\usage{
stat_boot_ci(
  mapping = NULL,
  data = NULL,
  geom = "ribbon",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  n = 1000,
  percent = 95,
  ...
)
}
\arguments{
\item{n}{number of resamples}

\item{percent}{percentile of credible interval}
}
\description{
Bootstrapping bivariate credible intervals for a ggplot
}
\keyword{additions,}
\keyword{data}
\keyword{ggplot}
\keyword{robust}
\keyword{statistics}
\keyword{vis,}
