% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whitening.R
\name{whitening}
\alias{whitening}
\title{Generates Mahalanobis spatial pre-whitening matrix.
Optionally regularizes matrix towards diagonal according to the Ledoit-Wolf method.
Depends: tawny package for Ledoit-Wolf method.}
\usage{
whitening(E, shrinkage = "ledoitwolf")
}
\arguments{
\item{E}{channel (e.g., vertex)-by-timepoint matrix of residuals.}

\item{shrinkage}{use Ledoit-Wolf method ("LW" or "ledoitwolf"), or manually specify lambda (e.g., 0.4).}
}
\description{
Generates Mahalanobis spatial pre-whitening matrix.
Optionally regularizes matrix towards diagonal according to the Ledoit-Wolf method.
Depends: tawny package for Ledoit-Wolf method.
}
