% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collate_surface_params.R
\name{collate_surface_params}
\alias{collate_surface_params}
\title{Read an AFNI gifti image and reshapes to matrix.}
\usage{
collate_surface_params(name, space = "hcp", xlabels = NULL, pattern = NULL)
}
\arguments{
\item{name}{name of .gii file to read}

\item{space}{the surface space the .gii is in. \%in\% c("hcp", "fsave")}

\item{xlabels}{optional; a character vector that indicates the exact strings of the sub-bricks to be extracted. If omitted, all sub-bricks are extracted.}
}
\description{
Optionally extracts sub-bricks by labels.
}
\keyword{AFNI,}
\keyword{gifti}
\keyword{neuroimaging,}
