% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collate_surface_params.R
\name{collate_surface_params}
\alias{collate_surface_params}
\title{Common vector math.}
\usage{
collate_surface_params(a)
}
\arguments{
\item{v}{a numeric vector}

\item{m}{a numeric vector or matrix}
}
\description{
Common vector math.
}
\keyword{algebra}
\keyword{centering,}
\keyword{linear}
\keyword{normalizing,}
\keyword{scaling,}
