% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_surface.R
\name{plot_surface}
\alias{plot_surface}
\title{Plot statistics on a surface.}
\usage{
plot_surface(
  overlay,
  underlay = hcp,
  hues = viridis::viridis(500),
  pos.max = NULL,
  pos.min = NULL,
  neg.max = NULL,
  neg.min = NULL,
  thresh0 = TRUE,
  facet.names = c("left_lateral", "left_medial", "right_lateral", "right_medial"),
  dims = c(1, 4),
  anno = FALSE,
  facet.params,
  anno.cex = 0.5,
  not.rstudio.gd = TRUE,
  ...
)
}
\description{
Plot statistics on a surface.
}
