% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_surface.R
\name{plot_surface}
\alias{plot_surface}
\title{Plot statistics on a surface.}
\usage{
plot_surface(
  overlay,
  underlay = hcp,
  hues = viridis::viridis(500),
  pos.max = NULL,
  pos.min = NULL,
  neg.max = NULL,
  neg.min = NULL,
  thresh0 = TRUE,
  facet.names = c("left_lateral", "left_medial", "right_lateral", "right_medial"),
  dims = c(1, 4),
  anno = FALSE,
  facet.params,
  anno.cex = 0.5,
  not.rstudio.gd = TRUE,
  ...
)
}
\arguments{
\item{overlay}{list containing values to plot on surface}

\item{underlay}{list containing underlay image}

\item{hues}{color pallete}

\item{thresh0}{set zero values to NA? (default == TRUE)}

\item{facet.names}{a vector of elements \%in\% c("left_lateral", "left_medial", "right_lateral", "right_medial"). Order determines order of facets within of plot grid.}

\item{dims}{the dimensions of the plot matrix to set up (prod(dims) must equal length(facet.names))}

\item{anno}{add text indicaing range of values displayed?}

\item{anno.cex}{size of annotation}

\item{not.rstudio.gd}{do not draw plot within rstudio (as rstudio device can be slow!); default is TRUE}

\item{pos.max/neg.min}{optional; values more extreme than these arguments will be set to pos.max/neg.min. If not specified (default), no thresholding occurs.}

\item{pos.min/neg.max}{optional; values less extreme these will be set to 0. If not specified (default), no thresholding occurs.}
}
\description{
Given an overlay (e.g., statistics) and underlay (e.g., group-template gifti file).
Overlay and underlay must have same structure: lists of length two, with names "L" and "R".
Elements within .$L and .$R correspond to vertices.
}
\details{
Much of this was adapted from Jo Etzel and John Muscheli.
See here: http://mvpa.blogspot.com/2018/06/tutorial-plotting-gifti-images-in-r.html
}
\keyword{AFNI,}
\keyword{data}
\keyword{gifti,}
\keyword{neuroimaging,}
\keyword{vis}
