% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_surface.R, R/plot_surfaces.R
\name{plot_surface}
\alias{plot_surface}
\title{Plot statistics on a surface.}
\usage{
plot_surface(
  overlay,
  underlay,
  facet.names = c("left_lateral", "left_medial", "right_lateral", "right_medial"),
  dims = c(1, 4),
  facet.params
)

plot_surface(
  overlay,
  underlay,
  facet.names = c("left_lateral", "left_medial", "right_lateral", "right_medial"),
  dims = c(1, 4),
  facet.params
)
}
\arguments{
\item{overlay}{values (colors) to plot; same structure as underlay}

\item{underlay}{a list of length 2, with elements titled c("L", "R"), corresponding to each hemisphere of a gifti of the desired underlay.}

\item{facet.names}{character vector with elements \%in\% c("left_lateral", "left_medial", "right_lateral", "right_medial"). Determines the row-wise ordering of the plot matrix.}

\item{dims}{Dimensions of plot matrix (row, col).}

\item{facet.params}{Optional list, of same structure as facet.names, containing theta and phi elements which specify orientations of each facet.}

\item{hues}{color pallete}

\item{pos.max/neg.min}{optional; values more extreme than these arguments will be set to pos.max/neg.min. If not specified (default), no thresholding occurs.}

\item{pos.min/neg.max}{optional; values less extreme these will be set to 0. If not specified (default), no thresholding occurs.}

\item{thresh0}{set zero values to NA? (default == TRUE)}

\item{anno}{add text indicaing range of values displayed?}

\item{anno.cex}{size of annotation}

\item{not.rstudio.gd}{do not draw plot within rstudio (as rstudio device can be slow!); default is TRUE}
}
\description{
Given an overlay (e.g., statistics) and underlay (e.g., group-template gifti file).
Overlay and underlay must have same structure: lists of length two, with names "L" and "R".
Elements within .$L and .$R correspond to vertices.

Much of this was adapted from code written by Joset Etzel
(http://mvpa.blogspot.com/2018/06/tutorial-plotting-gifti-images-in-r.html) and by John Muscheli (from library(gifti)).
}
\details{
Much of this was adapted from Jo Etzel and John Muscheli.
See here: http://mvpa.blogspot.com/2018/06/tutorial-plotting-gifti-images-in-r.html
}
\keyword{AFNI,}
\keyword{data}
\keyword{gifti}
\keyword{gifti,}
\keyword{neuroimaging,}
\keyword{vis}
